+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [1.4.2]
 
+### Added
+- Added activation dependency check for Linux and Windows
 
 ## [1.4.1]